+++ src/main/scala/com/redbubble/util/cache/CacheOps.scala
@@ -3,7 +3,8 @@ package com.redbubble.util.cache
 import java.util.concurrent.Executor
 
 import com.github.benmanes.caffeine.cache.Caffeine
-import com.redbubble.util.cache.redis.SimpleRedisCache
+import com.redbubble.util.cache.memory.{NonLoggingCaffeineCache, StatsCounter}
+import com.redbubble.util.cache.redis.MetricsEnableRedisCache
 import com.redbubble.util.metrics.StatsReceiver
 
 import scala.concurrent.duration.Duration
@@ -31,10 +32,20 @@ private[cache] object CacheOps {
     ScalaCache(NonLoggingCaffeineCache(underlying))
   }
 
-  // TODO Make use of the TTL?
+  /**
+    * Create a new in-memory cache, with metrics tracking.
+    *
+    * @param name          The name of the cache, used for sending metrics with the cache name.
+    * @param host          The hostname of the Redis instance.
+    * @param port          The port of the Redis instance.
+    * @param ttl           The time to live for items in the cache.
+    * @param executor      The executor to use when performing async cache operations.
+    * @param statsReceiver Where to log metrics to on the cache behaviour. Metrics are scoped by `name`.
+    */
   def newRedisCache(name: String, host: String, port: Int, ttl: Duration, executor: Executor)
       (implicit statsReceiver: StatsReceiver): ScalaCache[Array[Byte]] = {
-    val underlying = SimpleRedisCache(name, host, port)(executor, statsReceiver)
+    // TODO Make use of the TTL?
+    val underlying = MetricsEnableRedisCache(name, host, port)(executor, statsReceiver)
     ScalaCache[Array[Byte]](cache = underlying)
   }
